+++ src/fluoscenepy/__init__.py
@@ -3,10 +3,12 @@
 The "fluoscenepy" package is intended for simulation of a microscopic fluorescent image.
 
 @author: Sergei Klykov
-@licence: MIT, @year: 2024
+
+@licence: MIT, @year: 2025
+
 """
 
-__version__ = "0.0.3"  # Straightforward way of specifying package version and including it to the package attributes
+__version__ = "0.0.4"  # Straightforward way of specifying package version and including it to the package attributes
 
 if __name__ == "__main__":
     __all__ = ['fluoscene']  # for specifying 'from fluoscenepy import *' if package imported from some script